requirements.txt
gymnasium[all]
torch
numpy
einops
pyyaml
